fix: fps

+++ encord/orm/label_row.py
@@ -402,7 +402,7 @@ class LabelRowMetadataDTO(BaseDTO):
     is_shadow_data: bool = False
 
     """Only available for the VIDEO and AUDIO data_type"""
-    frames_per_second: Optional[int]
+    frames_per_second: Optional[float]
     number_of_frames: int
     duration: Optional[float]
